Failing experiments to include embassy as async executor for both web and native
For now only code for native, which fails.

It builds, but then upon opening the site, one gets an error on the initial import from "env".
This seems to be caused by a combination of the web wasm_bindgen target and some `unresolved import` from rust, likely caused by some mangling or not mangling from embassy.

Before this i tried to do this with the async main with attribute, as in the embassy wasm example, but then I got a compiler error that main is there twice.
Note that the embassy example is using the more usual library rust code approach.
However my experiments at moving towards that also failed just the same.